Refactor node type implementation
