+++ backend/cmd/server/repository/resources/graphs/auth_flow_config_github.json
@@ -4,7 +4,7 @@
     "nodes": [
         {
             "id": "git_auth",
-            "type": "TASK_EXECUTOR",
+            "type": "TASK_EXECUTION",
             "inputData": [
                 {
                     "name": "code",
