+++ backend/cmd/server/repository/resources/graphs/auth_flow_config_google.json
@@ -4,7 +4,7 @@
     "nodes": [
         {
             "id": "google_auth",
-            "type": "TASK_EXECUTOR",
+            "type": "TASK_EXECUTION",
             "inputData": [
                 {
                     "name": "code",
